requirements.txt
# requirements.txt
fastapi
uvicorn
python-dotenv
httpx
itsdangerous
python-multipart
mcp
